multimodal-summarizer/requirements.txt
streamlit
whisper
moviepy
gTTS
Pillow
langchain
langchain-groq
langchain-google-genai
google-generativeai
PyPDF2
pydub
